+++ SoftRend/render/Math3DStructs.h
@@ -116,9 +116,9 @@ namespace math3D
 			int x,y,z; // used for algorithm
 
 			// make sure values are all positive
-			x = fabs(this->x) * 1024;
-			y = fabs(this->y) * 1024;
-			z = fabs(this->z) * 1024;
+			x = static_cast<int>(fabs(this->x) * 1024.f);
+			y = static_cast<int>(fabs(this->y) * 1024.f);
+			z = static_cast<int>(fabs(this->z) * 1024.f);
 
 			// sort values
 			if (y < x)
